- To avoid bugs when running under x64 OS, several functions have changed their default "unsigned int" type. size_t is now used instead. - Lots of warnings and casts have been removed with this change. - Memory leak at HTTPAPI::GetPathFromLocationHeader() fixed

+++ HTTPCore/trunk/HTTPCore/CookieHandling.h
@@ -107,7 +107,7 @@ public:
 	int MatchCookie(char *lppath,char *lpdomain,char *lpCookieName, int securevalue);
 	void SetDate(time_t cExpire);
 	time_t GetDate(void) { return expire;}
-	int path_matches (const char *RequestedPath);
+	size_t path_matches (const char *RequestedPath);
 	char *GetCookieName(void) { return (lpCookieName); }
 	char *GetCookieValue(void) { return (lpCookieValue); }
 	char *GetCookiePath(void)  { return (path); }
